fix: remove unused line and struct and remove defer and rename processor/callback to processor/notification
