+++ internal/sbi/processor/processor.go
@@ -16,12 +16,6 @@ type Processor struct {
 	consumer *consumer.Consumer
 }
 
-type HandlerResponse struct {
-	Status  int
-	Headers map[string][]string
-	Body    interface{}
-}
-
 func NewProcessor(amf ProcessorAmf) (*Processor, error) {
 	p := &Processor{
 		ProcessorAmf: amf,
